Amiga beta 2
Upgrade JK2 and Ghoul2 code (dec 2014)

Fix texnum in R_CreateImage

+++ code/qcommon/msg.cpp
@@ -504,8 +504,9 @@ entityState_t communication
 
 typedef struct {
 	const char	*name;
-	int		offset;
-	int		bits;		// 0 = float
+	size_t		offset; // was int - Cowcat
+	int		bits;	// 0 = float
+
 } netField_t;
 
 // using the stringizing operator to save typing...
